+++ src/image.rs
@@ -16,6 +16,7 @@ use crate::datetime::datetime_from_unix_timestamp;
 #[cfg(feature = "chrono")]
 use chrono::{DateTime, Utc};
 
+#[derive(Debug)]
 /// Interface for accessing and manipulating a named docker image
 ///
 /// Api Reference: <https://docs.docker.com/engine/api/v1.41/#tag/Image>
@@ -93,6 +94,7 @@ impl<'docker> Image<'docker> {
     }
 }
 
+#[derive(Debug)]
 /// Interface for docker images
 pub struct Images<'docker> {
     docker: &'docker Docker,